+++ locales/it.js
@@ -1,0 +1,19 @@
+/* it - Italian Translation by anonymous ♥
+ Translation strings used by the bot
+ Strings wrapped in { } will be auto-replaced with updated values
+ Globals: {servername}, {invite}, (all from config) and {playercount} (current connected players)
+*/
+const locale = {
+    // Includes {name} + Globals
+    checkingWhitelist: "Ciao {name}, Stiamo verificando se possiedi la whitelist..",
+    discordNotOpen: "Devi aprire Discord prima di avviare FiveM (Riavvia entrambi se il problema persiste)",
+    fatalError: "Qualcosa è andato storto mentre stavamo verificando se possiedi la whitelist, per favore riprova tra un po' o contatta il supporto nel nostro discord se il problema persiste.",
+    notInDiscordServer: "Non sei nel Discord di {servername}, per favore entra: {invite}",
+    notWhitelisted: "Non hai il ruolo nel discord di {servername} richiesto per entrare nel server, possiedi la whitelist?",
+    kickedWithoutReason: "Sei stato cacciato dallo staff.",
+    announcement: "ANNUNCIO",
+    directMessage: "STAFF",
+    staffchat: "STAFFCHAT",
+};
+
+module.exports = locale;